+++ temba/channels/handlers.py
@@ -839,6 +839,9 @@ class InfobipHandler(BaseChannelHandler):
                 except iso8601.ParseError:
                     msg_date = None
 
+                if not text:
+                    continue
+
                 if channel.address.lstrip('+') == receiver.lstrip('+'):
                     urn = URN.from_tel(sender)
                     sms = Msg.create_incoming(channel, urn, text, date=msg_date, external_id=external_id)
@@ -1989,7 +1992,7 @@ class ChikkaHandler(BaseChannelHandler):
         action = self.get_param('message_type').lower()
 
         # look up the channel
-        channel = Channel.objects.filter(uuid=request_uuid, is_active=True, channel_type=Channel.TYPE_CHIKKA).exclude(org=None).first()
+        channel = Channel.objects.filter(uuid=request_uuid, is_active=True, channel_type='CK').exclude(org=None).first()
         if not channel:
             return HttpResponse("Error, channel not found for id: %s" % request_uuid, status=400)